+++ node_package/src/buildConsoleReplay.ts
@@ -1,6 +1,8 @@
 import RenderUtils from './RenderUtils';
 import scriptSanitizedVal from './scriptSanitizedVal';
 
+declare let debugConsole: any;
+
 declare global {
   interface Console {
     history?: {
@@ -21,6 +23,9 @@ export function consoleReplay(): string {
       let val;
       try {
         val = (typeof arg === 'string' || arg instanceof String) ? arg : JSON.stringify(arg);
+        if (val === undefined) {
+          val = 'undefined';
+        }
       } catch (e) {
         val = `${e.message}: ${arg}`;
       }