+++ node_package/src/serverRenderReactComponent.ts
@@ -10,10 +10,11 @@ import handleError from './handleError';
 import type { RenderParams } from './types/index';
 
 export default function serverRenderReactComponent(options: RenderParams): string {
-  const { name, domNodeId, trace, props, railsContext } = options;
+  const { name, domNodeId, trace, props, railsContext, throwJsErrors } = options;
 
   let htmlResult = '';
   let hasErrors = false;
+  let renderingError = null;
 
   try {
     const componentObj = ComponentRegistry.get(name);
@@ -68,19 +69,34 @@ Function Component.`);
       }
     }
   } catch (e) {
+    if (throwJsErrors) {
+      throw e;
+    }
+
     hasErrors = true;
     htmlResult = handleError({
       e,
       name,
       serverSide: true,
     });
+    renderingError = e;
   }
 
   const consoleReplayScript = buildConsoleReplay();
 
-  return JSON.stringify({
+  const result = {
     html: htmlResult,
     consoleReplayScript,
     hasErrors,
-  });
+  };
+
+  if (renderingError) {
+    result.renderingError = {
+      message: renderingError.message,
+      stack: renderingError.stack,
+    }
+  }
+
+
+  return JSON.stringify(result);
 }